+++ cudax/test/stf/places/managed.cu
@@ -33,12 +33,12 @@ __global__ void axpy(double a, slice<const double> x, slice<double> y)
   }
 }
 
-double X0(int i)
+double X0(size_t i)
 {
   return sin((double) i);
 }
 
-double Y0(int i)
+double Y0(size_t i)
 {
   return cos((double) i);
 }
